+++ packages/base/sync/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+* Wait until collections persistence was initialized before starting sync
+
 ## [1.0.1] - 2025-01-10
 
 ### Changed